Removed IConstraintResolver and moved ForsterConstraintResolver to order package, deleting its own package. Reasoning: There is (and probably always will be) only one ConstraintResolver.
